+++ internal/services/discord/discord.go
@@ -19,6 +19,12 @@ func NewDiscordService(token string) (*DiscordService, error) {
 
 	session.AddHandler(func(s *discordgo.Session, m *discordgo.Ready) {
 		fmt.Println("Connected to Discord as", m.User.Username)
+		fmt.Println("Invite URL: https://discordapp.com/oauth2/authorize?client_id=" + m.User.ID + "&scope=bot&permissions=0")
+		fmt.Println("Currently on", len(m.Guilds), "servers")
+		// Print out the servers we're on
+		for _, guild := range m.Guilds {
+			fmt.Println("  ", guild.Name, "(", guild.ID, ")")
+		}
 	})
 
 	return &DiscordService{session}, nil
@@ -32,4 +38,10 @@ func (d *DiscordService) Connect() error {
 // Disconnect disconnects from Discord
 func (d *DiscordService) Disconnect() error {
 	return d.session.Close()
+}
+
+// SendMessage sends a message to a channel
+func (d *DiscordService) SendMessage(channelID, message string) error {
+	_, err := d.session.ChannelMessageSend(channelID, message)
+	return err
 }
