+++ internal/weekbot/bot.go
@@ -2,24 +2,9 @@ package weekbot
 
 import (
 	"fmt"
-	"os"
-	"os/signal"
 
-	"weekbot/services/discord"
+	"weekbot/internal/services/discord"
 )
-
-func NewBot(config Config) *Bot {
-	discordService, err := discord.NewDiscordService(config.DiscordToken)
-	if err != nil {
-		fmt.Println("Error connecting to Discord:", err)
-		return nil
-	}
-
-	return &Bot{
-		config: &config,
-		dsc:    discordService,
-	}
-}
 
 // Bot is the main struct for the weekbot package
 type Bot struct {
@@ -27,10 +12,23 @@ type Bot struct {
 	dsc    *discord.DiscordService
 }
 
+func NewBot(config Config) (*Bot, error) {
+	discordService, err := discord.NewDiscordService(config.DiscordToken)
+	if err != nil {
+		fmt.Println("Error Configuring Discord Client:", err)
+		return nil, err
+	}
+
+	bot := &Bot{
+		config: &config,
+		dsc:    discordService,
+	}
+
+	return bot, nil
+}
+
 // Run starts the bot
 func (b *Bot) Run() {
-	defer b.dsc.Disconnect()
-
 	// Connect to Discord using the token from the config
 	fmt.Println("Connecting to Discord...")
 	err := b.dsc.Connect()
@@ -39,9 +37,9 @@ func (b *Bot) Run() {
 		return
 	}
 
-	stop := make(chan os.Signal, 1)
-	signal.Notify(stop, os.Interrupt)
-
-	<-stop
 	fmt.Println("Shutting down...")
+}
+
+func (b *Bot) Stop() {
+	b.dsc.Disconnect()
 }
